% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dimred.R
\name{theme_dimred}
\alias{theme_dimred}
\title{Minimal Theme for Dimension Reduction Plots}
\usage{
theme_dimred()
}
\value{
A \code{\link[ggplot2:theme]{ggplot2::theme()}} object.
}
\description{
A simple helper theme for dimension reduction (e.g., UMAP, PCA, t-SNE) plots.
It removes axis ticks and text and enforces a square aspect ratio.
}
\examples{
ggplot(iris, aes(Sepal.Length, Sepal.Width)) +
  geom_point() +
  theme_dimred()
}
